[refactor] DTO naming 통일
