+++ src/main/java/com/ll/quizzle/domain/point/dto/request/PointHistoryRequest.java
@@ -1,13 +1,13 @@
-package com.ll.quizzle.domain.point.dto;
+package com.ll.quizzle.domain.point.dto.request;
 
 import com.ll.quizzle.domain.point.type.PointType;
 
-public record PointHistoryRequestDTO(
+public record PointHistoryRequest(
 	PointType type, // "REWARD" or "USE", nullable
 	int page,
 	int size
 ) {
-	public PointHistoryRequestDTO {
+	public PointHistoryRequest {
 		// 기본값 설정
 		if (page < 0)
 			page = 0;
