+++ src/main/java/com/ll/quizzle/domain/point/dto/response/PointHistoryResponse.java
@@ -1,17 +1,17 @@
-package com.ll.quizzle.domain.point.dto;
+package com.ll.quizzle.domain.point.dto.response;
 
 import java.time.LocalDateTime;
 
 import com.ll.quizzle.domain.point.entity.Point;
 
-public record PointHistoryResponseDTO(
+public record PointHistoryResponse(
 	String type,
 	int amount,
 	String description,
 	LocalDateTime createdAt
 ) {
-	public static PointHistoryResponseDTO from(Point point) {
-		return new PointHistoryResponseDTO(
+	public static PointHistoryResponse from(Point point) {
+		return new PointHistoryResponse(
 			point.getType().name(),
 			point.getAmount(),
 			point.getReason().getDescription(),
